reuse modifiers for request params
also it's helpful to understand how we modify params
